- BUG: Data summaries were wrong when the number of alleles broke the 32bit INT limit.

+++ src/jhi/flapjack/gui/dialog/analysis/AlleleStatisticsDialog.java
@@ -20,7 +20,7 @@ public class AlleleStatisticsDialog extends JDialog implements ActionListener
 
 	private AlleleStatisticsPanelNB nbPanel;
 
-	public AlleleStatisticsDialog(GTViewSet viewSet, ArrayList<int[]> results)
+	public AlleleStatisticsDialog(GTViewSet viewSet, ArrayList<long[]> results)
 	{
 		super(
 			Flapjack.winMain,
